+++ docutils/HISTORY.rst
@@ -213,6 +213,7 @@ Release 0.22b.dev (unpublished)
   - Do not output .UR/.UE macros without refuri in node.
   - Use .MT/.ME macros for mailto.uris.
   - If macro references is active output refuri always.
+  - For text references: no line end after refuri.
 
 * docutils/writers/null.py